Renaming User and Product Controller names
